all tests created

+++ server/routes/statuses.js
@@ -12,10 +12,14 @@ export default (app) => {
       reply.render('statuses/index', { statuses });
       return reply;
     })
-    .get('/statuses/new', { name: 'getStatusPage', preValidation: app.authenticate }, (req, reply) => {
-      const status = new app.objection.models.status();
-      reply.render('statuses/new', { status });
-    })
+    .get(
+      '/statuses/new',
+      { name: 'getStatusPage', preValidation: app.authenticate },
+      (req, reply) => {
+        const status = new app.objection.models.status();
+        reply.render('statuses/new', { status });
+      },
+    )
     .get(
       '/statuses/:id/edit',
       { name: 'editStatus', preValidation: app.authenticate },
@@ -33,9 +37,6 @@ export default (app) => {
       { name: 'createStatus', preValidation: app.authenticate },
       async (req, reply) => {
         const status = req.body.data;
-        if (!status.id) {
-          status.id = uniqid();
-        }
         try {
           const validStatus = await app.objection.models.status.fromJson(status);
           await app.objection.models.status.query().insert(validStatus);
@@ -44,7 +45,7 @@ export default (app) => {
         } catch (error) {
           if (error instanceof ValidationError) {
             req.flash('error', i18next.t('flash.statuses.create.error'));
-            reply.render('statuses/new', { status: req.body.data, errors: error.data });
+            reply.render('statuses/new', { status, errors: error.data });
             return reply.code(422);
           }
           console.error(error);
@@ -57,19 +58,18 @@ export default (app) => {
       { name: 'updateStatus', preValidation: app.authenticate },
       async (req, reply) => {
         try {
-          const {
-            body: { data },
-          } = req;
-          // console.log('DATA: ', data);
-          const patchData = _.omit(data, 'id');
           const status = await app.objection.models.status.query().findById(req.params.id);
+          // console.log('ALL: ', await app.objection.models.status.query());
           // console.log('OLDDATA: ', status);
-          await status.$query().patch(patchData);
+          // console.log('NEWDATA: ', req.body.data);
+          await status.$query().patch(req.body.data);
+          // console.log('DATA UPDATED');
           req.flash('success', i18next.t('flash.statuses.edit.success'));
           reply.redirect(app.reverse('statuses'));
           return reply;
         } catch (error) {
           if (error instanceof ValidationError) {
+            // console.log('UPDATE ERROR: ', error);
             req.flash('error', i18next.t('flash.statuses.edit.error'));
             reply.render('statuses/edit', {
               status: { ...req.body.data, id: req.params.id },
@@ -86,13 +86,13 @@ export default (app) => {
       { name: 'deleteStatus', preValidation: app.authenticate },
       async (req, reply) => {
         const status = await app.objection.models.status.query().findById(req.params.id);
-        // const statusTasks = await status.$relatedQuery('tasks');
-        // if (statusTasks.length !== 0) {
-        //   req.flash('error', i18next.t('flash.statuses.delete.error'));
-        // } else {
-        await status.$query().delete();
-        req.flash('info', i18next.t('flash.statuses.delete.success'));
-        // }
+        const statusTasks = await status.$relatedQuery('tasks');
+        if (statusTasks.length !== 0) {
+          req.flash('error', i18next.t('flash.statuses.delete.error'));
+        } else {
+          await status.$query().delete();
+          req.flash('info', i18next.t('flash.statuses.delete.success'));
+        }
         reply.redirect(app.reverse('statuses'));
         return reply;
       },
